relocate function in keck_deimos
