+++ pypeit/spectrographs/keck_deimos.py
@@ -27,14 +27,11 @@ from pypeit import telescopes
 from pypeit import io
 from pypeit.core import parse
 from pypeit.core import framematch
-from pypeit.core import wave
-from pypeit import specobj, specobjs
 from pypeit.spectrographs import spectrograph
 from pypeit.images.detector_container import DetectorContainer
 from pypeit import dataPaths
 from pypeit.images.mosaic import Mosaic
 from pypeit.core.mosaic import build_image_mosaic_transform
-from pypeit import utils
 
 from pypeit.par import parset
 from pypeit.spectrographs import slitmask 
@@ -1926,87 +1923,87 @@ def deimos_read_1chip(hdu,chipno):
     return data, oscan
 
 
-def load_wmko_std_spectrum(fits_file:str, outfile=None, pad = False, split=True):
-    """Load up a Standard spectrum generated by WMKO IDL scripts
-    of the great Greg Wirth
-
-    The SpecObjs generated is checked that it is ready for fluxing
-
-    Args:
-        fits_file (str): filename
-        outfile ([type], optional): Write the SpecObjs object to a FITS file. Defaults to None.
-        pad (bool,optional): True if the resulting SpecObjs should be padded to an even length. Defaults to False
-        split (bool,optional): True if the resulting SpecObjs should be split into two detectors, False if it
-                               should be treated as a mosaic. Defaults to True.
-
-    Returns:
-        specobjs.SpecObjs: object holding the spectra
-    """
-
-    # Open up
-    hdul = fits.open(fits_file)
-    meta = Table(hdul[1].data)
-    idl_spec = Table(hdul[2].data)
-
-    # Hope this always works..
-    if split:
-        npix = int(len(idl_spec)/2)
-    else:
-        npix = len(idl_spec)
-
-    if pad:
-        if int(len(idl_spec)) % 2 != 0:
-            npix = int((len(idl_spec)+1)/2)
-            idl_spec.add_row([0.0, 0.0])
-
-    # Generate vacuum wavelengths
-    idl_vac = wave.airtovac(idl_spec['WAVELENGTH']*units.AA)
-
-    # Generate SpecObj
-    if not split:
-        sobj1 = specobj.SpecObj.from_arrays('MultiSlit', idl_vac.value[0:npix],
-                                            idl_spec['COUNTS'].data[0:npix],
-                                            1./(idl_spec['COUNTS'].data[0:npix]),
-                                            np.ones(idl_spec['COUNTS'].data[0:npix].size),
-                                            DET='MSC03')
-    else:
-        sobj1 = specobj.SpecObj.from_arrays('MultiSlit', idl_vac.value[0:npix],
-                                            idl_spec['COUNTS'].data[0:npix],
-                                            1./(idl_spec['COUNTS'].data[0:npix]),
-                                            np.ones(idl_spec['COUNTS'].data[0:npix].size),
-                                            DET='DET03')
-        
-        sobj2 = specobj.SpecObj.from_arrays('MultiSlit', idl_vac.value[npix:],
-                                            idl_spec['COUNTS'].data[npix:],
-                                            1./(idl_spec['COUNTS'].data[npix:]),
-                                            np.ones(idl_spec['COUNTS'].data[npix:].size),
-                                            DET='DET07')
-
-    # SpecObjs
-    sobjs = specobjs.SpecObjs()
-    sobjs.add_sobj(sobj1)
-    if split:
-        sobjs.add_sobj(sobj2)
-
-    # Fill in header
-    coord = utils.radec_to_coord((meta['RA'][0], meta['DEC'][0]))
-    sobjs.header = dict(EXPTIME=1., 
-                        AIRMASS=float(meta['AIRMASS'][0]), 
-                        DISPNAME=str(meta['GRATING'][0]), 
-                        SLMSKNAM='', # No slitmask is used for the throughput data
-                        PYP_SPEC='keck_deimos', 
-                        PYPELINE='MultiSlit',
-                        BINNING=f'{meta["SPEC_BIN"][0]},{meta["SPEC_BIN"][0]}',
-                        RA=coord.ra.deg, 
-                        DEC=coord.dec.deg
-                   )
-
-    # Check
-    assert sobjs.ready_for_fluxing()
-
-    # Write?
-    if outfile is not None:
-        sobjs.write_to_fits(sobjs.header, outfile)
-        print("Wrote: {}".format(outfile))
-
-    return sobjs
+# def load_wmko_std_spectrum(fits_file:str, outfile=None, pad = False, split=True):
+#     """Load up a Standard spectrum generated by WMKO IDL scripts
+#     of the great Greg Wirth
+# 
+#     The SpecObjs generated is checked that it is ready for fluxing
+# 
+#     Args:
+#         fits_file (str): filename
+#         outfile ([type], optional): Write the SpecObjs object to a FITS file. Defaults to None.
+#         pad (bool,optional): True if the resulting SpecObjs should be padded to an even length. Defaults to False
+#         split (bool,optional): True if the resulting SpecObjs should be split into two detectors, False if it
+#                                should be treated as a mosaic. Defaults to True.
+# 
+#     Returns:
+#         specobjs.SpecObjs: object holding the spectra
+#     """
+# 
+#     # Open up
+#     hdul = fits.open(fits_file)
+#     meta = Table(hdul[1].data)
+#     idl_spec = Table(hdul[2].data)
+# 
+#     # Hope this always works..
+#     if split:
+#         npix = int(len(idl_spec)/2)
+#     else:
+#         npix = len(idl_spec)
+# 
+#     if pad:
+#         if int(len(idl_spec)) % 2 != 0:
+#             npix = int((len(idl_spec)+1)/2)
+#             idl_spec.add_row([0.0, 0.0])
+# 
+#     # Generate vacuum wavelengths
+#     idl_vac = wave.airtovac(idl_spec['WAVELENGTH']*units.AA)
+# 
+#     # Generate SpecObj
+#     if not split:
+#         sobj1 = specobj.SpecObj.from_arrays('MultiSlit', idl_vac.value[0:npix],
+#                                             idl_spec['COUNTS'].data[0:npix],
+#                                             1./(idl_spec['COUNTS'].data[0:npix]),
+#                                             np.ones(idl_spec['COUNTS'].data[0:npix].size),
+#                                             DET='MSC03')
+#     else:
+#         sobj1 = specobj.SpecObj.from_arrays('MultiSlit', idl_vac.value[0:npix],
+#                                             idl_spec['COUNTS'].data[0:npix],
+#                                             1./(idl_spec['COUNTS'].data[0:npix]),
+#                                             np.ones(idl_spec['COUNTS'].data[0:npix].size),
+#                                             DET='DET03')
+#         
+#         sobj2 = specobj.SpecObj.from_arrays('MultiSlit', idl_vac.value[npix:],
+#                                             idl_spec['COUNTS'].data[npix:],
+#                                             1./(idl_spec['COUNTS'].data[npix:]),
+#                                             np.ones(idl_spec['COUNTS'].data[npix:].size),
+#                                             DET='DET07')
+# 
+#     # SpecObjs
+#     sobjs = specobjs.SpecObjs()
+#     sobjs.add_sobj(sobj1)
+#     if split:
+#         sobjs.add_sobj(sobj2)
+# 
+#     # Fill in header
+#     coord = utils.radec_to_coord((meta['RA'][0], meta['DEC'][0]))
+#     sobjs.header = dict(EXPTIME=1., 
+#                         AIRMASS=float(meta['AIRMASS'][0]), 
+#                         DISPNAME=str(meta['GRATING'][0]), 
+#                         SLMSKNAM='', # No slitmask is used for the throughput data
+#                         PYP_SPEC='keck_deimos', 
+#                         PYPELINE='MultiSlit',
+#                         BINNING=f'{meta["SPEC_BIN"][0]},{meta["SPEC_BIN"][0]}',
+#                         RA=coord.ra.deg, 
+#                         DEC=coord.dec.deg
+#                    )
+# 
+#     # Check
+#     assert sobjs.ready_for_fluxing()
+# 
+#     # Write?
+#     if outfile is not None:
+#         sobjs.write_to_fits(sobjs.header, outfile)
+#         print("Wrote: {}".format(outfile))
+# 
+#     return sobjs
